+++ clang/include/clang/AST/Stmt.h
@@ -179,7 +179,7 @@ protected:
 
     unsigned : NumExprBits;
 
-    unsigned Kind : 2;
+    unsigned Kind : 3;
     unsigned IsType : 1; // true if operand is a type, false if an expression.
   };